JUnit test project
The JUnit project now works fine. It is necessary to create a
Sampler_Testing project in the workspace where the JUnit is executed and
place the testing files in a tests folder in that project.

The JUnit performance test runs the configurated tests and save its
results in a generic folder. Finnally, it compacts the results.