JDTASTS JUnit Test
This update includes the JUnit files to execute the JDTAST performance
tests. It is necessary include in the JUnit Workspace a new Project
called Testing_JDTAST and, in a folder called "sets" add the sets of
data of JDTAST models.

The buttons in the SAMPLER plug-in to run the tests have been deleted.