Implemented Global Algorithm
Now it is possible use a new Abstraction Global Strategy: the Connected
Components Step. It is included in a isolated Algorithm, but it can be
used a a single step for custom algorithms.

This steps has a simple Configuration: a boolean indicating if the
components are strong or weak.

It is implemented using Tarjan Algorithm.