Spectral Algorithms
Spectral Algorithms and techniques have been included in a new plug-in
project. This has been included in the feature of SAMPLER.

It is needed a Linear Algebra library to make this algorithms work. It
is possible to use some numerical libraries too.

The spectral Project have LALG package with the interfaces needed by
Spectral Algorithms to work. It just needs a Matrix, a Vector and a
Eigenvector Dexomposition implementations.

Basic (but useless) implementations of Matrix and Vector has been given.