Two more Spectral Clustering Steps
Implemented two news Global Steps using the Spectral Clustering
Strategy:
- IterativeClusterBuilder: creates k clusters using the k th eigenvector
of the Laplacian matrix and then apply the iterative clustering with k/2
clusters to the positive and negative parts.
- HierarchicalClusterBuilder: creates 2 clusters using the BiCluster
strategy. Then take the lowest energy subgraph (the easiest cuttable)
and make there a new BiCluster until reach a total of k clusters.

There are no algorithm using this global steps, but it is possible
create a custom algorithm using these steps to give them a try.