New runnable version
This is a new version of SAMPLER.
- The Update Site has been updated
- The feature project has been updated too.

Features:
- Tha Layout System has been changed. Now a abstract interface is
provided to get a Graph layout. It is neccesary implement
LayoutAlgorithm interface and extend the layoutAlgorithms extension
point provided in the new plug-in rioko.layouts.
- Two example layouts are provided.
- The SAMPLER interface now use this new Layout system. The clases
ZestLayoutAlgorithm and ZestLayoutBridge allow to join Zest layouts with
our new system.
- Now it is possible open a SAMPLER editor with no model generated code.
It is neccessary to register the meta-model or have the ".ecore" file in
the same folder that the model we are opening.
- In rioko.plugin.utils plug-in, the Log system have change the target
folder. Now use the
${workspace.dir}/.metadata/.plugins/rioko.plugin.utils$ folder to save
the logs.
- The new editors opened from a ComposeNode have a custom name (not Zest
Editor as before) which refer to the first element of the compose node.
- The graph structure allows to mark the nodes.