Installable version of SAMPLER (not tested)
All the changes have been done. There is a new plug-in dedicated to
implement all the neccesary classes to open XMI files and its proxies.

The DrawModels Plug-In is now incapable to open files itself. It needs a
Reader implementation to give to a extension of the ModelDiagram class.

In the Utils plug-in a new class have been created: RegistryManagement.
This class has several method to get information of the Eclipse Registry
System and instantiate elements in a simpler way.

The Update Site have been updated.