Added functionality opening files
Now, when a file want to be openedwith the SAMPLER tool, and its file
extension is not registered in a Diagram Extension Point, Eclipse throws
a Dialog asking for a Reader to open the file.

If any error happens opening the file (i.e. a bad reader is selected) it
shows a dialog with the error.

If a file extension has multiple diagrams possible, now Eclipse shows an
information message and ask for a Reader among the extensions that have
that file extension registered.

The update Site has been updated.