Controlled a Concurrency Modfication Exception
The SAMPLER event gestor had a Concurrency Modification Exception (CME),
which makes the load of a new model fails.

It has been added a lock system on the Event/Listeners Map, so the CME
is avoided making the changes after the lock is over.

NEXT PROBLEM: when opening a new Zest Editor with the OpenFile handler,
the properties views are not properly linked with the editor, avoiding
to configurate correctly the Abstraction Algorithms and to modify the
properties of the Editor. It has been checked that the Porperties View
are linked when another Editor is selected and you come back to the
unlinked editor.