Fixed some bugs
We have fixed the bug mentiones in the last commits. Now, when a new
model is opened using the OpenFile handler, the configuration tables are
correctly linked to that editor.

We also have fixed another bug which provokes that the algorithm
configuration do not affect in copies of a ZestProperties because no
listener was generated.