Basic Structure for Saving Displays
The saving displays system has been changed and now it is based on a
simple class (ZestDisplay). It needs a IdParser from the diagram
extension point, and it is probably easy to get it.

NEXT STEP: Implement the UI part of this plug-in so we can test the
ZestDisplay class.