Basic Operating Save/Load System
Now it is possible (and works) saving a model viualization for each
model.
It is saved in a text file with "smp" extension.
The buttons on toolbar and meno don't let choose what display the user
wants to load, and when saving, always overwrite the same file. If
anyone wants to save more displays, he should change the name of the
"smp" file and load it using the rick click over it.