EMF-Splitter integration
Include a first attemp to integrate SAMPLER with the EMF-Splitter tool
(http://antoniogarmendia.github.io/EMF-Splitter/).

There is a project including a handler to open in the EMF-Splitter a
SAMPLER editor.

Updated all the Update Site references in the previous features.
Create a new feature for the new plugin.
Create a new Category on the Update Site to add all the features and
plugins that integrate EMF-Splitter with SAMPLER