Directory Drawing Integrated with EMF-SPlitter
Added a new project to integrate the Directory Drawer with EMF-Splitter
tool.
Now it is possible use the menu "Open With" of EMF-Splitter projects
with the Directory visualization, and, when using it, it is possible to
double-click a Resource and open the EMF Model Fragment associated with
that Resource.

Create the feature corresponding with that new functionality and added
to the update site on the category of EMF-Splitter integration.

Some changes have been done in the precious projects:
* Added a new extension point on project rioko.drawModels to declare
different listener over the GraphViewer of SAMPLER editor.
* Adapted the previous listeners so now they use the Extension Point